windows: fix inability to use local volume paths in ext2fs_open2
Local volume paths are Windows paths that start with "\\?\" such as
"\\?\Volume{12345678-1234-1234-1234-123456789ABC}\". They are one of
the most common paths one can use to designate a volume to format on
Windows. However, ext2fs_open2() is designed to interpret any question
mark in the provided path as an io_option which makes it unsuitable
to use such paths.

This patch ensures that we only process a question mark an an option
if it is not directly followed by backslash.

Signed-off-by: Pete Batard <pete@akeo.ie>